services/judge/src/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
grpcio>=1.60.0
grpcio-tools>=1.60.0
pybreaker==1.0.1
pydantic==2.5.3
pymongo==4.6.1
motor==3.3.2
protobuf>=5.26.1
pandas==2.1.4
numpy==1.26.2
joblib==1.3.2
